Rename single-word components to avoid confusion with html elements
