+++ bankid-idp/bankid-idp-frontend/src/components/AppFooter.vue
