+++ bankid-idp/bankid-idp-frontend/src/components/AppHeader.vue
