+++ bankid-idp/bankid-idp-frontend/src/components/StatusItem.vue
